seongwon component 폴더 생성, YoutubeInfo component 생성, BoardView YoutubeInfo 컴포넌트 사용
